replace dependency to double type with traact::Scalar

+++ src/traact/datatypes.h
@@ -8,7 +8,7 @@
 #include <traact/traact_core_export.h>
 namespace traact {
 
-using Scalar = double;
+using Scalar = float;
 
 using TimeDuration = std::chrono::duration<int64_t, std::nano>;
 using Timestamp = std::chrono::time_point<std::chrono::system_clock, TimeDuration>;
